backend/requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
firebase-admin
